update from Xcode 26.0.1 / iOS base SDK 26.0 to Xcode 26.4.1 / iOS base SDK 26.4 (#456)
the only immediate code change necessary is casting numeric enum values
to an explicit int, in places where arithmetic with floating point
values is done.

the change is necessary because a calculation without an explicit type
conversion may result in a compiler error in future C++ versions. the
potential future error is indicated by a new compiler warning generated
by the compiler that is shipped with Xcode 26.4.1.

+++ src/newgame/KomiSelectionController.m
@@ -218,7 +218,7 @@ enum NoHandicapGameSectionItem
     }
     case NoHandicapGameSection:
     {
-      komiValue = MinimumKomiItem + (indexPath.row / 2.0);
+      komiValue = (int)MinimumKomiItem + (indexPath.row / 2.0);
       assert(komiValue >= MinimumKomiItem && komiValue <= MaximumKomiItem);
       break;
     }
@@ -250,7 +250,7 @@ enum NoHandicapGameSectionItem
   else
   {
     section = NoHandicapGameSection;
-    row = 2 * (komiValue - MinimumKomiItem);
+    row = 2 * (komiValue - (int)MinimumKomiItem);
     assert(komiValue >= MinimumKomiItem && komiValue <= MaximumKomiItem);
   }
   return [NSIndexPath indexPathForRow:row inSection:section];
